email-ai-agent/requirements.txt
docx==0.2.4
extract_msg==0.54.1
Flask==3.1.1
openai==1.84.0
Pillow==11.2.1
protobuf==6.31.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.0
python_docx==1.1.2
Werkzeug==3.1.3
